+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## (next)
+## 2.0.0-beta.1
 
 > Deprecated `AsyncElevatedButton`, `AsyncTextButton` and `AsyncOutlinedButton`